+++ client/coral-ui/index.js
@@ -3,5 +3,5 @@ export {default as CoralLogo} from './components/CoralLogo';
 export {default as FabButton} from './components/FabButton';
 export {default as TabBar} from './components/TabBar';
 export {default as Tab} from './components/Tab';
-export {default as ContentBox} from './components/ContentBox';
+export {default as TabContent} from './components/TabContent';
 export {default as Button} from './components/Button';
